better Cursor::seek
the original algorithm started at one end and moved towards the click until the distance to the click increased, which made it prone to being 'caught' on fractions when the distance increases when moving from denominator to numerator.

the new algorithm instead moves towards the click until it's on the other side of the click.