fix occasional white blocks in selection in IE<9
Since IE<9 doesn't support CSS transforms, we use the matrix
filter instead [1], but that won't correctly antialias unless
the background is an opaque solid color. So the background is
set to normally be white, but when in a selection the color
should be the background highlight color. However, under
certain conditions [2], IE<9 won't realize it should be
matching different CSS rules and the background remains
white surrounded by blue, so we force IE to re-check what
rules match by assigning to .className.

[1]: https://github.com/laughinghan/mathquill/wiki/Transforms
[2]: It seems to be if an element already in the DOM, such as
a sqrt, is added to another element already in the DOM, such
as a selection. The sqrt now matches a different CSS rule
that says any descendant of a selection, even matrix filtered
elements, should now have a blue background, but IE doesn't
notice.