call redraw only once per node, in post-order depth-first #optimization
The `.find(':not(:has(:first)))` was a nontrivial ~20ms in IE8 with a
couple parens and brackets, but the real performance obscenity was
calling .redraw() on every ancestor of every leaf node, which always
involve costly DOM measurements and often even involve exorbitantly
expensive 2D matrix transformations.

And post-order depth-first traversal is almost trivial thanks to
MathElement::eachChild() and good ol' recursion.