skipTextInput is not needed.
When we changed from the virtual keypress event to the virtual textinput
event, we started checking whether text was actually entered.  This
supersedes the need to set a flag.

The flag was necessary when the virtual keypress event handlers were
trying to tell whether text was entered based on `which`, and couldn't,
except based on information obtained during keydown.  In particular,
arrow keys have the same `which` on keypress as &, %, etc.