split Parser::then into then, map, and result
Parser::then takes either a parser or a function that returns a parser.
It will throw an error if a parser is not returned.

Parser::map takes a function that transforms the output of the parser it
is called on.

Parser::result takes a specific value to return from the parser,
regardless of the output of previous parsers.