public MathQuill functions throw, w/ good messages
when jQuery resolves the argument to more or less than a single HTML
element, or MathQuill.MathField etc is called on a MathQuill instance
of a different kind, we throw an exception with an appropriate error
message.

The idea is, exceptions are good for failing fast when something is
very, very wrong. Guaranteeing that MathQuill is only ever passed a
single element is pretty easy, so not doing so, or trying to make the
same HTML element two different kinds of MathQuill instances, probably
means something is very, very wrong.

Aside from these two situations, MathQuill generally tries to make do
with whatever, such as MathQuill-ifying something that has already
been MathQuill-ified (to the same type) being a no-op.