add Node::dispose() method to free cached nodes
If references to nodes aren't deleted from the cache, there will be a
memory leak as no Node object nor any object or closure directly or
indirectly referenced by any Node object can ever be garbage collected.
This is a core action on Nodes and should always have been a method,
defined next to the initializer that set the reference from the cache in
the first place.

This is basically manual memory management, the [dispose pattern][] is
essentially C++'s [Resource Acquisition Is Initialization][RAII] idiom
that was cutting edge in the 1980's, adapted for environments with
automatic garbage collection, which sucks, but we're pretty much stuck
with it until [weak maps][] arrive in ECMAScript Harmony.

[dispose pattern]: http://en.wikipedia.org/wiki/Dispose_pattern
[RAII]: http://en.wikipedia.org/wiki/Resource_Acquisition_Is_Initialization
[weak maps]: http://wiki.ecmascript.org/doku.php?id=harmony:weak_maps