add and use variadic() util HOF
make the function that send() returns variadic and use it in
Node::bubble and Node::postOrder, so that they're both variadic and can
take in either a string (which will be assumed to be a method name) or a
function

send() now also fails silently, since it's being used by bubble and
postOrder. Note that this changes the semantics of bubble slightly, so
that if a method name maps to a falsy value on a Node, bubble used to
ignore it but now will attempt to call it, like postOrder would've.
This never actually happens, though, so it's all good.