fix opaque matrixed el's in IE8, eps. nthroots
since IE8 doesn't support CSS transforms, the Matrix filter is used to
stretch square roots etc, which antialiases wrong when the background
isn't opaque, so we set them all to have an opaque background.

The problem is that, in particular, the [n] of nthroots gets obscured by
the opaque background of the square root radical.

After much research and trial and error, I've ended up using the Chroma
filter as the only way to make the background transparent. This is
plagued with problems, documented in [the wiki page on Transforms][1].

[1]: https://github.com/mathquill/mathquill/wiki/Transforms