Cursor::withDirInsertAt only blurs if parent changes
Fixes keyboard selection throwing exceptions.

This is so TextPiece::selectTowards (which calls Cursor::insertAdjacent
which calls Cursor::withDirInsertAt) won't cause
TextBlock::consolidateChildren to be called (by TextBlock::blur), which
messes up the tree positions that the cursor and anticursor are at when
doing selection with the keyboard.

Also, don't bother waiting til after moving the cursor to call
this.parent.blur(), that was some bullshit I pulled to hack together the
original TextBlock:
https://github.com/mathquill/mathquill/commit/e15a85fffbb3c1f0cdbdd1bf1c3fe4ee05581e7c#L2L44

Note that this only narrowly fixes keyboard selection throwing
exceptions, focus/blur are still fucked; if you blur by clicking on the
address bar, for example, then tab back, the cursor's gone and the
tree's messed up.