cursor.parent.keystroke() replaces bubbling 'onKey'
since 'keystroke' isn't bubbled, it now has to be on Node.prototype
instead of just on the root. Don't want the huge switch/case in tree.js,
so it goes in the newly created interaction.js, the idea is eventually
all the browser events stuff in rootelements.js will be moved there.

also since 'keystroke' isn't bubbled, LatexCommandInput::onKey has to be
moved to a .keystroke on the MathBlock inside the LatexCommandInput.

when overriding keystroke, instead of return false to block bubbling,
have to explicitly call _super.keystroke to continue default.