fix \class whitelist regex syntax error #ImAnIdiot
apparently octal escapes are actually illegal in ES5 [1], but browsers
are nice and accept them...but only for two characters? MDN [2] wasn't
specific, but that's the behavior I got in Chrome, at least. So the
syntax error in Chrome was that '0' to '\037' was an invalid range
(since the higher end was lower than the lower end).

[1]: http://es5.github.io/#x15.10.2.11 , "NOTE"
[2]: https://developer.mozilla.org/en-US/docs/JavaScript/Guide/Regular_Expressions#Using_Special_Characters
     search for "octal"