Don't stopPropagation() mouse events
...instead, use closest root block.

Completely replace the old RootMathCommand-descended, jQadd-overriding
\editable. Instead, LatexCmds.editable::finalizeTree sets up its block
to be like RootMathBlock after .mathquill('editable').

The main reason for the big createRoot() refactor, 46a97ee "Split
createRoot into many functions", is that the mousedown handlers attached
to the root of the static math and to the root of the inner editable
stepped on each other's toes. Until now we did stopPropagation, but
that's bad (it breaks e.g. the Desmos graphing calculator), the right
thing is for mousedown to be a delegated event, with the handler only
attached to the ultimate root of the MathQuill tree (the root of the
static math), which then uses the closest ancestor that's a root block
to determine editability, attaches the mousemove handler to it, focuses
its textarea, and seeks its cursor.

Note that it's important to attach the mousemove event handler to the
closest ancestor root block, lest the target node not descend from the
cursor's root, breaking Cursor::seek. (The mousemove handler attached
to the document ignores the target node.)