Move general latex im/export {rootelements => latex}.js
(By "general" I mean not specific to particular commands.)

I left the code that, on a 'copy' event, gets the selection LaTeX, in
rootelements.js for now just try to strip down Root{Math,Text}Block.

I know it looks like I just took some code that branched with
polymorphism and turned it into code that branches on a flag, but that's
only because...that's exactly what I did. But the point is to centralize
the branching in publicapi.js, so there's eventually one place to split
up into a class hierarchy that branches on polymorphism.

Also had to change the up/down unit test, which kinda reaches in and
helps itself to the root block methods, but hey, there's one less place
that relies root.cursor!