Improve cursor edge case after .write('^2')
For example, if the cursor is just before a superscript:

    x|^n

and you .write('^2'), they get "contact welded":

    x^{2n}

but the cursor should still be right of the thing just typed, and left
of the stuff originally to its right:

    x^{2|n}

contactWeld now takes an optional cursor, which will be checked if the
above special case applies and needs to be put anywhere special.
Only Controller::writeLatex passes in the optional cursor (to
MathElement::finalizeInsert, which passes it straight through to
.postOrder('contactWeld', cursor). MathCommand::createLeftOf previously
passed in an vestigal, ignored cursor to .finalizeInsert(), now it is
important that createLeftOf NOT pass in a cursor, because createLeftOf
calls .placeCursor()).