.sibling{Created,Deleted}() supersedes .respace() #rearchitect #doTodo
.respace() was the only hook called on a very useful event, the creation
or deletion of an immediate sibling, but wasn't provided enough
information. Replace .respace() with hooks with specific information
about the creation or deletion of immediate siblings.

This lets Variable, which wants to auto-unitalicize contiguous runs of
letters whenever they're edited, to only search for contiguous runs once
per insertion or deletion of a letter, rather than each letter
.respace() is called on (that's 3 letters for a letter inserted between
2 other letters) repeating the same search.

Note that neither .siblingCreated() nor .siblingDeleted() are called on
creation/insertion, but .contactWeld() is and the .jQ and siblings are
guaranteed to have been set up. (Is this abuse? I was originally
thinking .finalizeTree() but latex.test.js needs to call .finalizeTree()
for .latex() on SupSub's to work, but .jQ isn't set up so
PlusMinus::finalizeTree() blows up.)

Finally, get rid of Node::respace = noop, all the places where it saves
having to do feature detection, have to do a null check anyway.