Get rid of unitalicized fns containing 'f'
They're unimportant, current LatexCmds.f doesn't work with
auto-unitalicization because it descends from Symbol but not Variable,
and it's a hassle to get it to work because the way the visual
unitalicization works is to change the font from Times New Roman to
Symbola, but unlike Roman letters the florin character is still in
italic style, so 'hcf' looks like an unitalicized 'hc' but an italicized
'f'.

This should go away when MathQuill finally switches to STIX fonts, or
when I finally ragedelete LatexCmds.f.