Delete duplicate .siblingDeleted()'s in unwrapGramp #cleanup
Cursor::unwrapGramp is only called by MathBlock::deleteOutOf which is
only called by Controller::deleteDir (src/services/keystroke.js), which
then immediately calls .siblingDeleted(), so Cursor::unwrapGramp doing
so is an unnecessary duplicate call.

This duplicate call (to .siblingDeleted()'s predecessor, .respace()) has
been in there since Cursor::unwrapGramp was first written:
02e80e15826d7278ab82c283d2dd2e612b8283ee
 (it was called Cursor::unwrapParent then and you can see that it ends
  with respace calls, but the callers, Cursor::backspace/deleteForward,
  call respace immediately after calling `this.unwrapParent()`)

Interesting architectural note:

This was a layer violation: the hooks are high-level node methods and
should only be called by the controller or other high-level node
methods, not cursor methods, which should all be low-level tree and
manipulating-the-cursor-as-a-Point methods. Not coincidentally, it was
also blocking the next change I want to make, which is to change the
contract of .siblingDeleted() from .siblingDeleted(dir) to
.siblingDeleted(dir, deletedSibling).

Now, Controller::deleteDir() doesn't easily know what sibling was
deleted, because it delegates deletion to .delete{Towards,OutOf}(),
which can delete as much or as little as the node desires, so the
contract of .delete{Towards,OutOf}() has to change so the return value
tells Controller::deleteDir() the deleted nodes to pass to
.siblingDeleted().

By contrast, Cursor::unwrapGramp() knows what node it deleted and hence
*looks* like it would know what to pass to as the deleted sibling, but
actually the .delete{Towards,OutOf}() that called it could delete more
nodes and put the cursor anywhere. Cursor::unwrapGramp() *can't* know
what's going to happen, because it's in the layer *below*
.delete{Towards,OutOf}().