Deleting brackets reverts them to HalfBracket's
The .delete{Towards,OutOf}() architecture is so great! But not the
having to monkey-patch methods on the sole MathBlock in finalizeTree,
that sucks.

Also, having to add {Node,Fragment}::withDirAdopt() was annoying.
Not sure what could be done about that; it's non-trivial to
directionalize Fragment::adopt, because it relies on the order .each()
iterates through the nodes.