Allow mis-matched parens/brackets like [a, b)
Store on each Bracket the characters and control sequences of each side,
and swap them out when closing a one-sided bracket by typing a bracket
facing the opposite way. Also, when deleting one of a mis-matched pair
of brackets, the ghost is always a matching opposite of the remaining
side of the bracket.

Parser now calls Bracket constructor with arguments based on the
characters immediately after each of \left and \right, like it should,
rather than always a matched pair based on the character immediately
after \left.

All the tests that try deleting the parens from various permutations of
an expression with a paren pair in it, change to deleting a mis-matched
bracket pair. It's more general, I'm not really worried about something
working for a mis-matched pair but not a matching one, whereas I am
about the reverse.

Resolves mathquill/mathquill#150.